fixed type mismatch

+++ applications/samples/007_swapchain/source.cpp
@@ -270,7 +270,7 @@ llri::Device* createDevice(llri::Instance* instance, llri::Adapter* adapter)
         adapter,
         enabledFeatures,
         extensions.size(), extensions.data(),
-        queues.size(), queues.data()
+        static_cast<uint32_t>(queues.size()), queues.data()
     };
 
     llri::Device* device;
